fix: build

+++ service/persist/asset.go
@@ -46,8 +46,8 @@ type ErrAssetNotFoundByIdentifiers struct {
 
 // AssetRepository represents a repository for interacting with persisted contracts
 type AssetRepository interface {
-	GetByOwner(context.Context, EthereumAddress, Chain, int, int) ([]Asset, error)
-	GetByToken(context.Context, EthereumAddress, Chain, int, int) ([]Asset, error)
+	GetByOwner(context.Context, EthereumAddress, Chain, int64, int64) ([]Asset, error)
+	GetByToken(context.Context, EthereumAddress, Chain, int64, int64) ([]Asset, error)
 	GetByIdentifiers(context.Context, EthereumAddress, EthereumAddress, Chain) (Asset, error)
 	UpsertByIdentifiers(context.Context, EthereumAddress, EthereumAddress, Asset) error
 	UpdateByID(context.Context, DBID, interface{}) error
